# py/map: Add board option MICROPY_PY_MAP_ORDERED to make all dicts/maps ordered.