No change to normal builds, but if `MICROPY_PY_MAP_ORDERED` is set in port/board config then the ordered flag is set for all new maps (`dicts`)

While I presume this has a performance impact it is useful behavior for some applications eg. https://github.com/micropython/micropython/issues/5322