@dpgeorge 2019-11-13
Hmm, interesting.  I guess this really goes with the CPython 3.7 change that `dict` guarantees insertion order; https://docs.python.org/3/whatsnew/3.7.html

Although there is a subtly that maps are not 1:1 with dict, map is used to implement dict.  Not sure if that really matters though.

> While I presume this has a performance impact

Would be interesting to test it.  It would also mean that most of the code in `mp_map_lookup()` is now unused.