@andrewleech 2019-11-19
I've just expanded the #define's to chop out the hash based code that's no longer used if ordered is enabled. 
It does make the #defines's a little harder to visually follow, but on my build it reduces flash size by 1232 bytes.